requirements.txt
asgiref==3.4.1
Django==3.2.5
fpdf==1.7.2
python-docx==0.8.11
pytz==2021.1
sqlparse==0.4.1
docxtpl